requirements.txt
google-cloud-firestore>=0.32
flask
slackclient